MainScreen, UpdateScreen, CreateScreen, Navigation
Files for all screens and navigation created, only MainScreen and navigation implemented.

MainScreen:
Displays note titles in a top-to-bottom list. Titles are clickable and expands to show the rest of the contents, a delete button and a button that redirects to the UpdateScreen.

Also contains a floatingActionButton that redirects to CreateScreen