NoteData.kt, Theme.kt, Typography.kt
For clarification:

Earlier commits weren't added.

NoteData:
Is created to create input fields for text and title. Contrains variable to store input errors.

Renders a save-button. Uses the validateText and validateTitle functions from note.kt when clicked.

Theme:
Sets  up a theme, since the background is changed to white a theme for darkmode is not needed.

Looks fine on my device on darkmode.

Typography:
Adds a custom typograpthy.